pluggable conditions and effects
